Create an `IQuickWidget` (#190627)
To do this, I've pulled `QuickInputController` into its own file since `quickInput.ts` was getting too big...

And then I made another thing that extends `QuickInput` that hides pretty much everything but does allow for a title bar.

This could be useful for more "QuickWebView" type experiences if we wanna light anything like that up.
