+++ src/vs/platform/quickinput/common/quickInput.ts
@@ -214,13 +214,18 @@ export interface IQuickInput extends IDisposable {
 	description: string | undefined;
 
 	/**
-	 * Should be an HTMLElement (TODO: move this entire file into browser)
+	 * Should be an HTMLElement.
+	 * @deprecated Use an IQuickWidget instead
 	 */
 	widget: any | undefined;
 
 	step: number | undefined;
 
 	totalSteps: number | undefined;
+
+	buttons: ReadonlyArray<IQuickInputButton>;
+
+	readonly onDidTriggerButton: Event<IQuickInputButton>;
 
 	enabled: boolean;
 
@@ -233,6 +238,15 @@ export interface IQuickInput extends IDisposable {
 	show(): void;
 
 	hide(): void;
+
+	didHide(reason?: QuickInputHideReason): void;
+}
+
+export interface IQuickWidget extends IQuickInput {
+	/**
+	 * Should be an HTMLElement (TODO: move this entire file into browser)
+	 */
+	widget: any | undefined;
 }
 
 export interface IQuickPickWillAcceptEvent {
@@ -296,10 +310,6 @@ export interface IQuickPick<T extends IQuickPickItem> extends IQuickInput {
 	customLabel: string | undefined;
 
 	customHover: string | undefined;
-
-	buttons: ReadonlyArray<IQuickInputButton>;
-
-	readonly onDidTriggerButton: Event<IQuickInputButton>;
 
 	readonly onDidTriggerItemButton: Event<IQuickPickItemButtonEvent<T>>;
 
@@ -412,16 +422,6 @@ export interface IInputBox extends IQuickInput {
 	 * Event called when the user submits the input.
 	 */
 	readonly onDidAccept: Event<void>;
-
-	/**
-	 * Buttons to show in addition to user input submission.
-	 */
-	buttons: ReadonlyArray<IQuickInputButton>;
-
-	/**
-	 * Event called when a button is selected.
-	 */
-	readonly onDidTriggerButton: Event<IQuickInputButton>;
 
 	/**
 	 * Text show below the input box.
@@ -554,9 +554,14 @@ export interface IQuickInputService {
 	createQuickPick<T extends IQuickPickItem>(): IQuickPick<T>;
 
 	/**
-	 * Provides raw access to the quick input controller.
+	 * Provides raw access to the input box controller.
 	 */
 	createInputBox(): IInputBox;
+
+	/**
+	 * Provides raw access to the quick widget controller.
+	 */
+	createQuickWidget(): IQuickWidget;
 
 	/**
 	 * Moves focus into quick input.
